+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 Yii2 multiple input change log
 ==============================
 
-1.2.16 in development
+1.2.17 in development
 ---------------------
+
+1.2.16
+------
 
 - Bug #70: replacing of the placeholder after preparing the content of row